+++ unliftio/src/UnliftIO.hs
@@ -4,12 +4,10 @@ module UnliftIO
   ( module Control.Monad.IO.Unlift
   , module UnliftIO.Async
   , module UnliftIO.Chan
-  , module UnliftIO.Environment
   , module UnliftIO.Exception
   , module UnliftIO.IO
   , module UnliftIO.IORef
   , module UnliftIO.MVar
-  , module UnliftIO.Process
   , module UnliftIO.STM
   , module UnliftIO.Temporary
   , module UnliftIO.Timeout
@@ -18,12 +16,10 @@ module UnliftIO
 import Control.Monad.IO.Unlift
 import UnliftIO.Async
 import UnliftIO.Chan
-import UnliftIO.Environment
 import UnliftIO.Exception
 import UnliftIO.IO
 import UnliftIO.IORef
 import UnliftIO.MVar
-import UnliftIO.Process
 import UnliftIO.STM
 import UnliftIO.Temporary
 import UnliftIO.Timeout